Add files via upload
Create a game called minesweeper.

Create a function that takes a grid of # and -, where each hash (#) represents a
mine and each dash (-) represents a mine-free spot.

Return a grid, where each dash is replaced by a digit, indicating the number of
mines immediately adjacent to the spot i.e. (horizontally, vertically, and
diagonally)

Example of an input:
[ ["-", "-", "-", "#", "#"],
["-", "#", "-", "-", "-"],
["-", "-", "#", "-", "-"],
["-", "#", "#", "-", "-"],
["-", "-", "-", "-", "-"] ]

Example of the expected output:
[ ["1", "1", "2", "#", "#"],
["1", "#", "3", "3", "2"],
["2", "4", "#", "2", "0"],
["1", "#", "#", "2", "0"],
["1", "2", "2", "1", "0"] ]

Here is a tip. When checking adjacent positions to a specific position in the grid,
the following table might assist you in determining adjacent indexes:

    NW position       = (-1, -1) >> current_row: -1, current_col: -1
    N position        = (-1, 0) >> current_row: - 1, current_col: 0
    NE position       = (-1, 1) >> current_row: -1, current_col: +1
    W position        = (0, -1) >> current_row: 0, current_col: -1
    CURRENT POSITION  = (0, 0) >> current_row, current_col
    E position        = (0, 1) >> current_row: 0, current_col: +1
    SW position       = (1, -1) >> current_row: +1, current_col: -1
    S position        = (1, 0) >> current_row: +1, current_col: 0
    SE position       = (1, 1) >> current_row: +1, current_col: +1

Also, ensure that when checking adjacent positions in the grid you take into
account that on the edges of the grid, you may go out of bounds